+++ src/utils/parseTokens/parseToken/classInstruction/classInstruction.test.ts
