+++ src/utils/parseTokens/parseToken/classInstruction/classInstruction.ts
@@ -1,4 +1,4 @@
-import { State } from '../../parseTokens2'
+import { State } from '../../parseTokens'
 
 export function classInstruction() {
 	return (state: State) => {
