+++ src/utils/parseTokens/parseToken/idInstruction/idInstruction.ts
@@ -1,4 +1,4 @@
-import { State } from '../../parseTokens2'
+import { State } from '../../parseTokens'
 
 export function idInstruction() {
 	return (state: State) => {
