+++ src/utils/parseTokens/parseToken/multiplyInstruction/multiplyInstruction.test.ts
