+++ src/utils/parseTokens/parseToken/multiplyStartInstruction/multiplyStartInstruction.test.ts
