+++ src/utils/parseTokens/parseToken/nameInstruction/nameInstruction.test.ts
