+++ src/utils/parseTokens/parseToken/upInstruction/upInstruction.test.ts
